+++ src/main/groovy/nebula/plugin/ospackage/application/OspackageApplicationExtension.groovy
@@ -1,0 +1,5 @@
+package nebula.plugin.ospackage.application
+
+class OspackageApplicationExtension {
+    String prefix
+}