+++ src/main/groovy/nebula/plugin/ospackage/daemon/DaemonTemplateTask.groovy
@@ -1,14 +1,13 @@
 package nebula.plugin.ospackage.daemon
 
-import org.gradle.api.DefaultTask
+import org.gradle.api.internal.ConventionTask
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.OutputFiles
 import org.gradle.api.tasks.TaskAction
 
 /**
  * Monster class that does everything.
  */
-class DaemonTemplateTask extends DefaultTask {
+class DaemonTemplateTask extends ConventionTask {
     //@Input
     Map<String, String> context
 
@@ -20,9 +19,9 @@ class DaemonTemplateTask extends DefaultTask {
 
     @TaskAction
     def template() {
-        TemplateHelper templateHelper = new TemplateHelper(destDir, '/nebula/plugin/ospackage/daemon')
-        templates.collect{ String templateName ->
-            templateHelper.generateFile(templateName, context)
+        TemplateHelper templateHelper = new TemplateHelper(getDestDir(), '/nebula/plugin/ospackage/daemon')
+        getTemplates().collect { String templateName ->
+            templateHelper.generateFile(templateName, getContext())
         }
     }