+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-1.9.0 / 2014-xx-xx
+1.12.2
+==================
+
+* Mate together the application plugin with the ospackage plugin
+* Create daemon scripts for applications plugin projects
+ 
+1.12.0
+==================
+
+* Daemon tools scripts
+
+1.12.0
 ==================
 
 * Initial release